fix handle tag when update
